Split Structure.GetStructureCost() into two methods: GetStructureCost() and GetStructureCostByType(). Also split Town.CanCreateStructure() by adding CanCreateStructureByType(). Modified Town._resources to now take hold floats rather than ints, allowing for gradual increases over many frames. Added Town.Calculate() to calculate all necessary resource/structure modifications for a given frame. Tested with simple resource count increases. Added Town.TakeStructureCost() which now reduces the number of resources when a structure is built. Modified BuildStateUI to only display integer values (since the resources are stored as floats now). GhostStructure class now properly implements the cost of a structure when converting into a regular Structure.

+++ CityBuilder/Interface/BuildStateUI.cs
@@ -41,7 +41,7 @@ namespace CityBuilder.Interface
             }
         }
 
-        public bool UpdateResourceCounts(Dictionary<Resource.ResourceType, int> resources)
+        public bool UpdateResourceCounts(Dictionary<Resource.ResourceType, float> resources)
         {
             bool error = false;
             if (resources == null)
@@ -50,7 +50,7 @@ namespace CityBuilder.Interface
             {
                 if (_resourceTextBoxes.ContainsKey(type))
                 {
-                    int value = resources[type];
+                    int value = (int)resources[type];
                     if (value > Config.MAX_RESOURCE_VALUE)
                         value = Config.MAX_RESOURCE_VALUE;
                     if (!resourcesCounts.ContainsKey(type))
